fix(ios): Move track button inline with report issue button
- Create TrackTrainInlineButton component matching FeedbackButton style
- Place track button next to report issue in stops section
- Remove standalone TrainFollowPrompt card for cleaner UX
- Remove dead code (TrainFollowPrompt, PulsingModifier)
